+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Remove Poise dependencies
+- Rewrite resources as Chef custom resources
+
 ## 4.8.0 - *2021-09-28*
 
 - Added `advertise_addr_wan_ipv4` and `advertise_addr_wan_ipv6` attributes
