+++ metadata.rb
@@ -14,9 +14,4 @@ supports 'debian', '>= 9.0'
 supports 'ubuntu', '>= 16.04'
 supports 'windows'
 
-depends 'build-essential', '>= 5.0.0' # cookstyle: disable ChefModernize/UnnecessaryDependsChef14
 depends 'nssm', '>= 4.0.0'
-depends 'golang'
-depends 'poise', '~> 2.2'
-depends 'poise-archive', '~> 1.3'
-depends 'poise-service', '~> 1.4'
