--- templates/default/systemd.service.erb
@@ -1,16 +1,0 @@
-[Unit]
-Description=<%= @name %>
-Wants=network.target
-After=network.target
-
-[Service]
-Environment=<%= @environment.map {|key, val| %Q{"#{key}=#{val}"} }.join(' ') %>
-ExecStart=<%= @command %>
-ExecReload=/bin/kill -<%= @reload_signal %> $MAINPID
-KillSignal=<%= @stop_signal %>
-User=<%= @user %>
-WorkingDirectory=<%= @directory %>
-<%= "LimitNOFILE=#{node['consul']['service_nofile']}" if node['consul']['service_nofile'] %>
-
-[Install]
-WantedBy=multi-user.target
